requirements/dev.txt
django-browser-reload==1.3.0
-r requirements/production.txt
-r requirements/no-cache.txt